requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
discord.py==2.3.2
Flask==3.0.2
frozenlist==1.4.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.4
oauthlib==3.2.2
praw==7.7.1
prawcore==2.4.0
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==1.3.1
update-checker==0.18.0
urllib3==2.2.1
websocket-client==1.7.0
Werkzeug==3.0.1
yarl==1.9.4
